Add files via upload
Code sample, permission to share granted.

Combine VCFs intelligently combines multiple VCF (Variant Call Format) files
of the same, similar or different strain. E.g. Combine VCFs can be called on
several VCF files generated from different mice of the same Regeneron
proprietary strain in order to obtain a consensus "true" VCF.

Combine VCF takes in a list of VCF files (and optionally, a configuration TSV
file specifying criteria for filtering intersecting records, else the criteria 
is generated automatically). The VCF files are combined via a series of calls
upon Bcftools (newer versions only) resulting in a main consensus VCF file. 
Additionally, the number of records in relevant file intersections is noted and 
optionally the user may choose to keep certain intersection-VCF files created to
further construct the consensus VCF.

The combination step consists of obtaining all VCF records meeting the criteria
of sufficient quality with respect to both the number of VCF files the record
was found in and the total number of VCF files being combined. I.e. records that
have high quality (as measured by quantities such as Depth and Phred Quality
Depth) and appear in more VCFs are consequently more likely to be true positives
and are thus included in the consensus VCF, whereas records that appear sparsely
and/or have lower read qualities are more likely to have been false positives
and are thus not included in the consensus VCF.

E.g. the automatically generated criteria for records that appear in exactly 2 
out of 3 VCFs would be "DP>=24 && QD>=30.0" (depth/number of reads must exceed
24 and the quality-depth must exceed 30.0") and for a records that appear in 3 
out of 5 VCFs, the criteria is "DP>=20 && QD>=18.0."

Combine VCFs is intended to be ran only on occasion with typically 3 to 5 VCF
files, higher amounts are expected. Runtime averages around ~15 minutes on the 
Sun Grid System cluster and though can push 40 minutes for larger combinations.